feat(eidos): promote EvalFinding to eidos::knowledge::finding
Consolidates the three incompatible finding shapes (EvalFinding,
CheckResult, ProsocheResult) by making eidos own the canonical type.
This PR does the relocation only; wiring prosoche + self_audit to
emit Finding is a follow-up.

Back-compat alias left at eval::stats::finding so existing callers
continue to compile with zero churn. Research rationale:
/tmp/research-3511-quantified-self.md section 4 R1.

Closes #3779

Gate-Passed: kanon 0.1.0
